Add dry-struct integration via an extension
