+++ lib/dry/validation/extensions.rb
@@ -1,3 +1,7 @@
 Dry::Validation.register_extension(:monads) do
   require 'dry/validation/extensions/monads'
 end
+
+Dry::Validation.register_extension(:struct) do
+  require 'dry/validation/extensions/struct'
+end
