+++ lib/dry/validation/schema/class_interface.rb
@@ -87,10 +87,6 @@ module Dry
         klass =
           if other.is_a?(self)
             Class.new(other.class)
-          elsif other.is_a?(Class) && other < Dry::Struct
-            Validation.Schema(parent: target, build: false) do
-              other.schema.each { |attr, type| required(attr).filled(type) }
-            end
           elsif other.respond_to?(:schema) && other.schema.is_a?(self)
             Class.new(other.schema.class)
           else
