+++ lib/dry/validation/schema/value.rb
@@ -158,8 +158,6 @@ module Dry
             [type, [name, predicate(input, args)]]
           elsif input.respond_to?(:rule)
             [type, [name, [:type, input]]]
-          elsif input.is_a?(::Class) && input < ::Dry::Struct
-            [type, [name, [:schema, Schema.create_class(self, input)]]]
           elsif input.is_a?(Schema)
             [type, [name, schema(input).to_ast]]
           else
